refactor: Update AlbumService to use ConfigService for album folder path
This commit refactors the AlbumService in the src/modules/album/album.service.ts file to use the ConfigService to retrieve the album folder path. The album folder path is now stored in the private albumFolder property of the AlbumService class. This change improves the maintainability and flexibility of the code by centralizing the configuration of the album folder path.

The update includes the following changes:
- Import the ConfigService from the @nestjs/config package.
- Add a private albumFolder property to store the album folder path.
- Inject the ConfigService into the AlbumService constructor.
- Retrieve the album folder path from the ConfigService using the get method.
- Assign the retrieved album folder path to the albumFolder property.

This commit does not introduce any functional changes to the behavior of the AlbumService.